requirements.txt
# Updated requirements.txt
flask==2.3.2
python-dotenv==1.0.0
azure-storage-file-datalake==12.13.0
azure-data-tables==12.4.2
numpy==1.24.4
pandas==1.5.3
flask-cors==3.0.10
gunicorn==21.2.0
scipy==1.11.4
openai==1.55.3